requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
chromadb==0.5.5
sentence-transformers==3.1.1
pypdf==5.1.0
python-docx==1.1.2
groq==0.11.0
posthog>=2.4.0,<6.0.0
docx2txt==0.8
striprtf==0.0.26
httpx==0.27.2
